+++ ep-maven-parent/ep-maven/src/test/java/com/tibco/ep/buildmavenplugin/PackageTest.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (C) 2018-2025 Cloud Software Group, Inc.
+ * Copyright (C) 2018-2025. Cloud Software Group, Inc.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -445,7 +445,6 @@ public class PackageTest extends BetterAbstractMojoTestCase {
             }
             zipFile.close();
         } catch (IOException e) {
-            // TODO Auto-generated catch block
             e.printStackTrace();
         }